# [et] Speed up expotools launch time