--- a/tools/expotools/package.json
+++ b/tools/expotools/package.json
@@ -26,7 +26,7 @@
     "@expo/commander": "2.21.1",
     "@expo/json-file": "^8.2.6",
     "@expo/spawn-async": "^1.5.0",
-    "@expo/xdl": "~57.9.14",
+    "@expo/xdl": "~57.9.18",
     "@types/base-64": "^0.1.2",
     "@types/concat-stream": "^1.6.0",
     "@types/connect": "^3.4.32",