--- a/tools/expotools/tsconfig.json
+++ b/tools/expotools/tsconfig.json
@@ -15,6 +15,9 @@
     "noFallthroughCasesInSwitch": true,
     "esModuleInterop": true,
     "resolveJsonModule": true,
-    "skipLibCheck": true
+    "skipLibCheck": true,
+    "skipDefaultLibCheck": true,
+    "incremental": true,
+    "tsBuildInfoFile": "./cache/.tsbuildinfo"
   }
 }