# Why

As a number of expotools command is constantly growing the time has come to improve expotools launch time.

# How

- Instead of calculating checksum of source files and yarn.lock together, now their checksums are independent of each other. Thanks to this we can skip running `yarn` process in many cases.
- Each time TypeScript files need to be rebuilt, expotools wrapper will save a schema of commands so that on next run it will know which command file should be `require`d when that command is used – until now it's been importing all command files thus actually all files and dependencies (even those unused in the invoked command).
- Turned on TypeScript's incremental builds.

# Test Plan

## Previously

### Cold run (warm yarn + rebuilding)
```
et ppt -h      9.04s user 0.67s system 146% cpu 6.642 total
et publish -h  8.97s user 0.67s system 147% cpu 6.517 total
et urn -h      9.06s user 0.67s system 149% cpu 6.519 total
```

### Warm run
```
et ppt -h      0.64s user 0.13s system 112% cpu 0.685 total
et publish -h  0.65s user 0.14s system 112% cpu 0.702 total
et urn -h      0.64s user 0.13s system 113% cpu 0.684 total
```

## Now

### Cold run (warm yarn + rebuilding)
```
et ppt -h      6.76s user 0.58s system 137% cpu 5.326 total
et publish -h  6.69s user 0.55s system 144% cpu 5.023 total
et urn -h      6.62s user 0.55s system 144% cpu 4.953 total
```

### Cold run (only rebuilding)
```
et ppt -h      5.87s user 0.49s system 138% cpu 4.579 total
et publish -h  5.81s user 0.47s system 146% cpu 4.284 total
et urn -h      5.65s user 0.46s system 146% cpu 4.181 total
```

### Warm run
```
et ppt -h      0.48s user 0.10s system 113% cpu 0.512 total
et publish -h  0.29s user 0.07s system 112% cpu 0.314 total
et urn -h      0.16s user 0.04s system 108% cpu 0.179 total
```